Added MCMC main loop to refactor (TODO: currently not running correctly)
